requirements.txt
Django~=3.1.6
freeze==3.0
Pillow==8.1.0
pytz==2020.1
six==1.15.0
djangorestframework==3.11.0